split writers and loaders
