--- modules/builder/src/glb-loader/index.js
@@ -1,13 +1,0 @@
-// Public methods
-export {parseGLB} from './glb-loader';
-export {encodeGLB} from './glb-writer';
-
-// Experimental exports, exposes internals
-export {default as _GLBDecoder} from './glb-decoder';
-export {default as _GLBEncoder} from './glb-encoder';
-export {default as _GLBBufferPacker} from './glb-buffer-packer';
-export {default as _unpackGLBBuffers} from './unpack-glb-buffers';
-export {
-  packJsonArrays as _packJsonArrays,
-  unpackJsonArrays as _unpackJsonArrays
-} from './pack-json-arrays';
