+++ modules/builder/src/index.js
@@ -1,22 +1,10 @@
-// LOADERS
+// UTILS
+export {loadUri} from './utils/load-uri.js';
 
-export {loadUri} from './loader-utils/load-uri.js';
+// WRITERS
+export {encodeGLB, _GLBEncoder, _GLBBufferPacker, _packJsonArrays} from './writers/glb-writer';
+export {encodeBinaryXVIZ, XVIZWriter} from './writers/xviz-writer';
 
-export {parseBinaryXVIZ} from './xviz-loader/xviz-binary-loader';
-export {encodeBinaryXVIZ} from './xviz-loader/xviz-binary-writer';
-export {default as XVIZWriter} from './xviz-loader/xviz-writer';
-export {default as XVIZBuilder} from './xviz-loader/xviz-builder';
-export {default as XVIZMetadataBuilder} from './xviz-loader/xviz-metadata-builder';
-
-export {parseGLB} from './glb-loader/glb-loader';
-export {encodeGLB} from './glb-loader/glb-writer';
-
-// Experimental exports, exposes internals
-export {default as _GLBDecoder} from './glb-loader/glb-decoder';
-export {default as _GLBEncoder} from './glb-loader/glb-encoder';
-export {default as _GLBBufferPacker} from './glb-loader/glb-buffer-packer';
-export {default as _unpackGLBBuffers} from './glb-loader/unpack-glb-buffers';
-export {
-  packJsonArrays as _packJsonArrays,
-  unpackJsonArrays as _unpackJsonArrays
-} from './glb-loader/pack-json-arrays';
+// BUILDERS
+export {default as XVIZBuilder} from './builders/xviz-builder';
+export {default as XVIZMetadataBuilder} from './builders/xviz-metadata-builder';
