+++ modules/builder/src/utils/flatten.js
