+++ modules/builder/src/utils/get-image-size.js
