+++ modules/builder/src/utils/text-encoder-decoder.js
