+++ modules/builder/src/utils/to-low-precision.js
