+++ modules/builder/src/writers/glb-writer/index.js
@@ -1,0 +1,7 @@
+// Public methods
+export {encodeGLB} from './glb-writer';
+
+// Experimental exports, exposes internals
+export {default as _GLBEncoder} from './glb-encoder';
+export {default as _GLBBufferPacker} from './glb-buffer-packer';
+export {default as _packJsonArrays} from './pack-json-arrays';
