+++ modules/builder/src/writers/xviz-writer/index.js
@@ -1,3 +1,3 @@
 // Public methods
-export {parseBinaryXVIZ} from './xviz-binary-loader';
 export {encodeBinaryXVIZ} from './xviz-binary-writer';
+export {default as XVIZWriter} from './xviz-writer';
