+++ modules/builder/src/writers/xviz-writer/xviz-pack-binary.js
@@ -1,6 +1,6 @@
-import {default as GLBEncoder} from '../glb-loader/glb-encoder';
-import {GLBBufferPacker} from '../glb-loader';
-import {packJsonArrays} from '../glb-loader/pack-json-arrays';
+import GLBEncoder from '../glb-writer/glb-encoder';
+import GLBBufferPacker from '../glb-writer/glb-buffer-packer';
+import packJsonArrays from '../glb-writer/pack-json-arrays';
 
 const MAGIC_XVIZ = 0x5856495a; // XVIZ in Big-Endian ASCII
