+++ modules/parser/src/loaders/glb-loader/glb-decoder.js
@@ -1,6 +1,7 @@
 /* eslint-disable camelcase, max-statements */
-import {padTo4Bytes, TextDecoder} from '../loader-utils'; // Node.js polyfills
-import assert from '../loader-utils/assert';
+import {padTo4Bytes} from '../../utils/array-utils';
+import {TextDecoder} from '../../utils/text-decoder';
+import assert from '../../utils/assert';
 
 // glTF CONSTANTS
