+++ modules/parser/src/loaders/glb-loader/glb-loader.js
@@ -1,6 +1,6 @@
 import GLBDecoder from './glb-decoder';
 import unpackGLBBuffers from './unpack-glb-buffers';
-import {unpackJsonArrays} from './pack-json-arrays';
+import unpackJsonArrays from './unpack-json-arrays';
 
 export function parseGLB(arrayBuffer, options = {}) {
   const {json, binaryByteOffset} = GLBDecoder.parseGlbBuffer(arrayBuffer, options);
