+++ modules/parser/src/loaders/glb-loader/index.js
@@ -1,0 +1,7 @@
+// Public methods
+export {parseGLB} from './glb-loader';
+
+// Experimental exports, exposes internals
+export {default as _GLBDecoder} from './glb-decoder';
+export {default as _unpackGLBBuffers} from './unpack-glb-buffers';
+export {default as _unpackJsonArrays} from './unpack-json-arrays';
