+++ modules/parser/src/loaders/glb-loader/unpack-glb-buffers.js
@@ -1,4 +1,4 @@
-import assert from '../loader-utils/assert';
+import assert from '../../utils/assert';
 
 const TYPE_COMPONENTS = {
   SCALAR: 1,
