+++ modules/parser/src/loaders/xviz-loader/index.js
@@ -1,0 +1,2 @@
+// Public methods
+export {parseBinaryXVIZ} from './xviz-binary-loader';
