+++ modules/parser/src/utils/array-utils.js
@@ -1,0 +1,3 @@
+export function padTo4Bytes(byteLength) {
+  return (byteLength + 3) & ~3;
+}
